+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Main
+
+* Allow control of the graph state from the user code.
+* Deprecate `run` in favor of `run_with_callback` and `run_with_mpsc`.
+
 ## 1.0.1
 
 * Bump imgui version to 0.6.